requirements.txt
requests>=2.31.0
numpy>=1.24.0
pytz>=2024.1
python-dotenv>=1.0.0
anthropic>=0.40.0
schedule>=1.2.0

# Daily report sync (reports/daily_report_sync.py)
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
gspread>=6.0.0

# Floor elasticity model (intelligence/floor_model.py)
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
joblib>=1.3.0
